requirements.txt
hbutils>=0.9.0
pillow
numpy
scikit-learn
huggingface_hub
tqdm
opencv-python
opencv-contrib-python
pandas
scipy
emoji>=2.5.0
pilmoji>=1.3.0
shapely
pyclipper
deprecation>=2.0.0
hfutils>=0.2.3
filelock